rename to operator and add S3 update for tree bom
